# doc/openapi.yaml
openapi: 3.1.0
info:
  title: Homenavi External API
  version: 0.5.16
  description: |
    HTTP API exposed through the Homenavi API gateway. This document covers the
    supported public control-plane contract. Event streams are described in
    doc/asyncapi.yaml.
servers:
  - url: /
    description: Current deployment ingress origin
  - url: http://localhost
    description: Local Compose ingress
security:
  - bearerAuth: []
...
paths:
...
  /api/automation/workflows:
    get:
      tags: [Automation]
      operationId: listWorkflows
      responses:
        '200':
          description: Saved workflows
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorkflowList' }
    post:
      tags: [Automation]
      operationId: createWorkflow
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkflowInput' }
      responses: { '201': { $ref: '#/components/responses/JSONSuccess' } }
  /api/automation/workflows/{id}:
    get:
      tags: [Automation]
      operationId: getWorkflow
      parameters: [ { $ref: '#/components/parameters/WorkflowID' } ]
      responses: { '200': { $ref: '#/components/responses/JSONSuccess' } }
    put:
      tags: [Automation]
      operationId: updateWorkflow
      parameters: [ { $ref: '#/components/parameters/WorkflowID' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkflowInput' }
      responses: { '200': { $ref: '#/components/responses/JSONSuccess' } }
    delete:
      tags: [Automation]
      operationId: deleteWorkflow
      parameters: [ { $ref: '#/components/parameters/WorkflowID' } ]
      responses: { '200': { $ref: '#/components/responses/JSONSuccess' } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT sent as an Authorization bearer token. Browser clients may also use the auth cookie.
  parameters:
...
    WorkflowID: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
    JSONSuccess:
      description: Successful JSON response
      content: { application/json: { schema: { type: object, additionalProperties: true } } }
...
  schemas:
...
    WorkflowInput:
      type: object
      required: [name, definition]
      properties:
        name: { type: string }
        enabled: { type: boolean }
        definition: { type: object, additionalProperties: true }
        source: { type: object, additionalProperties: true }
    WorkflowList:
      type: object
      required: [workflows]
      properties:
        workflows: { type: array, items: { type: object, additionalProperties: true } }